+++ packages/beff-wasm/fixtures/codegen-snaps/decoder/bff-generated/validators.js
@@ -2,6 +2,7 @@
 /* eslint-disable */
 
 
+
 function pushPath(ctx, path) {
   if (ctx.paths == null) {
     ctx.paths = [];