+++ packages/beff-hono/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @beff/hono
 
+## 0.0.10
+
+### Patch Changes
+
+- Add ts nocheck to out files
+- Updated dependencies
+  - @beff/cli@0.0.10
+  - @beff/client@0.0.10
+
 ## 0.0.9
 
 ### Patch Changes