+++ packages/beff-react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @beff/react
 
+## 0.0.10
+
+### Patch Changes
+
+- Add ts nocheck to out files
+- Updated dependencies
+  - @beff/client@0.0.10
+
 ## 0.0.9
 
 ### Patch Changes