+++ packages/web-diff/CHANGELOG.md
@@ -1,5 +1,14 @@
 # web-diff
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @beff/cli@0.0.10
+  - @beff/hono@0.0.10
+  - @beff/react@0.0.10
+
 ## 0.1.1
 
 ### Patch Changes