+++ packages/worker-diff/CHANGELOG.md
@@ -1,5 +1,13 @@
 # worker-diff
 
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @beff/cli@0.0.10
+  - @beff/hono@0.0.10
+
 ## 0.0.1
 
 ### Patch Changes